+++ lib/src/portal/location/xdg_location_accuracy.dart
@@ -1,0 +1,2 @@
+/// Requested accuracy of location information.
+enum XdgLocationAccuracy { none, country, city, neighborhood, street, exact }
