+++ lib/src/portal/notification/xdg_notification_button.dart
@@ -1,0 +1,10 @@
+/// A button to be shown in a notification.
+class XdgNotificationButton {
+  /// Label on this button.
+  final String label;
+
+  /// Action to perform with this button.
+  final String action;
+
+  XdgNotificationButton({required this.label, required this.action});
+}
