+++ lib/src/portal/notification/xdg_notification_icon.dart
@@ -1,0 +1,36 @@
+import 'dart:typed_data';
+
+/// An icon to be shown in a notification.
+abstract class XdgNotificationIcon {}
+
+/// An icon stored in the file system.
+class XdgNotificationIconFile extends XdgNotificationIcon {
+  /// Path of this icon
+  final String path;
+
+  XdgNotificationIconFile(this.path);
+}
+
+/// An icon at a URI.
+class XdgNotificationIconUri extends XdgNotificationIcon {
+  /// Uri of this icon
+  final String uri;
+
+  XdgNotificationIconUri(this.uri);
+}
+
+/// A themed icon.
+class XdgNotificationIconThemed extends XdgNotificationIcon {
+  /// Theme names to lookup for this icon in order of priority.
+  final List<String> names;
+
+  XdgNotificationIconThemed(this.names);
+}
+
+/// An icon with image data.
+class XdgNotificationIconData extends XdgNotificationIcon {
+  /// Image data for this icon.
+  final Uint8List data;
+
+  XdgNotificationIconData(this.data);
+}
