+++ lib/src/portal/notification/xdg_notification_priority.dart
@@ -1,0 +1,2 @@
+/// Priorities for notifications.
+enum XdgNotificationPriority { low, normal, high, urgent }
