+++ lib/src/portal/xdg_portal_response.dart
@@ -1,0 +1,2 @@
+/// Response from a portal request.
+enum XdgPortalResponse { success, cancelled, other }
