+++ lib/src/portal/xdg_portals.dart
@@ -1,0 +1,7 @@
+export 'email/xdg_email_portal.dart';
+export 'location/xdg_location_portal.dart';
+export 'network/xdg_network_monitor_portal.dart';
+export 'notification/xdg_notification_portal.dart';
+export 'open_uri/xdg_open_uri_portal.dart';
+export 'proxy_resolver/xdg_proxy_resolver_portal.dart';
+export 'settings/xdg_settings_portal.dart';
